refactor: split app initialization from server connection
